move mgenepy under depmap_omics_upload
