+++ depmap_omics_upload/upload.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 import pandas as pd
 from datetime import date
 
-from mgenepy.utils import helper as h
+from depmap_omics_upload.mgenepy.utils import helper as h
 from depmap_omics_upload import tracker as track
 from taigapy import TaigaClient
 import json
